+++ src/main/java/co/edu/corposucre/productionfood/usuario/dto/RolResponse.java
@@ -1,3 +1,3 @@
 package co.edu.corposucre.productionfood.usuario.dto;
 
-public record RolResponse(Long idRol, String nombre, String descripcion) {}
+public record RolResponse(Integer idRol, String nombre, String descripcion) {}
